% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafflehouseindex-package.R
\docType{package}
\name{wafflehouseindex}
\alias{wafflehouseindex}
\alias{wafflehouseindex-package}
\title{Compute the Waffle House Index Value Based on Current Closed Location Data}
\description{
The Waffle House Index was one of a series of metrics used by a former FEMA directory to gauge
the severity of disasters. If a Waffle House closes, you know you're hosed. Tools are provided to retrieve
closure information and compute the Waffle House Index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/wafflehouseindex}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/wafflehouseindex/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
