% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-index.R
\name{whi_compute}
\alias{whi_compute}
\title{Compute the Waffle House Index from list of open/closed stores}
\usage{
whi_compute(whi = whi_get_loc_info())
}
\arguments{
\item{whi}{pass in a pre-gathered list of Waffle House stores (with \code{\link[=whi_get_loc_info]{whi_get_loc_info()}}) or let it compute it for you.}
}
\description{
Compute the Waffle House Index from list of open/closed stores
}
