% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info.R
\name{whi_get_loc_info}
\alias{whi_get_loc_info}
\title{Retrieve the current info on all Waffle House locations}
\usage{
whi_get_loc_info()
}
\description{
Retrieve the current info on all Waffle House locations
}
\note{
this function is memoised since you're hitting a hidden API 196 times.
}
